requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.29.3
botocore==1.32.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecs-logging==2.1.0
elastic-apm==6.19.0
fastapi==0.103.2
h11==0.14.0
idna==3.4
ifaddr==0.2.0
jmespath==1.0.1
Pillow==10.0.1
psutil==5.9.6
py-eureka-client==0.11.8
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.0
uvicorn==0.23.2
wrapt==1.14.1
